Moved ServerThread to Server.java
